pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = ["tokusumi <tksmtoms@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.60.1"
python-jose = "^3.1.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pytest-cov = "^2.10.0"
flake8 = "^3.8.3"
mypy = "^0.782"
black = "^19.10b0"
uvicorn = "^0.11.8"
botocore = "^1.17.32"
boto3 = "^1.14.32"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
